Big Refactor To Increase Readability and Maintainability
Most importantly split up Transcription Service
