+++ ForensicWhisperDeskZH/Properties/Settings.settings
@@ -1,9 +1,5 @@
 ﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="ForensicWhisperDeskZH.Properties" GeneratedClassName="Settings">
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)">
   <Profiles />
-  <Settings>
-    <Setting Name="ServerURI" Type="System.String" Scope="Application">
-      <Value Profile="(Default)">localhost:8000/</Value>
-    </Setting>
-  </Settings>
+  <Settings />
 </SettingsFile>
