+++ ForensicWhisperDeskZH/Transcription/WhisperModelManager.cs
@@ -22,179 +22,6 @@ namespace ForensicWhisperDeskZH.Transcription
         }
 
         /// <summary>
-        /// This method downloads all Available Whisper models and adds them to the Models Directory.
-        /// It Should only ever be called in a debug environment and not in production.
-        /// </summary>
-        public async Task DownloadAllModelsAsync()
-        {
-            // Create Models directory relative to the project/application directory
-            string projectDirectory = AppDomain.CurrentDomain.BaseDirectory;
-            string modelsDirectory = Path.Combine(projectDirectory, "Models");
-            //modelsDirectory = modelsDirectory.Replace("bin\\Debug\\", "");
-
-            // Ensure the Models directory exists
-            Directory.CreateDirectory(modelsDirectory);
-
-            // List of all available model types from GetGgmlType method
-            string[] modelTypes = { "tiny", "base", "small", "medium", "large", "turbo" };
-
-            var downloader = new WhisperGgmlDownloader(_httpClient);
-
-            foreach (string modelType in modelTypes)
-            {
-                try
-                {
-                    string fileName = $"ggml-{modelType}.bin";
-                    string modelPath = Path.Combine(modelsDirectory, fileName);
-
-                    // Skip if model already exists
-                    if (File.Exists(modelPath))
-                    {
-                        Console.WriteLine($"Model {modelType} already exists at {modelPath}");
-                        continue;
-                    }
-
-                    Console.WriteLine($"Downloading {modelType} model...");
-
-                    // Download the model
-                    using (var modelStream = await downloader.GetGgmlModelAsync(GetGgmlType(modelType)))
-                    using (var fileStream = File.Create(modelPath))
-                    {
-                        await modelStream.CopyToAsync(fileStream);
-                    }
-
-                    Console.WriteLine($"Successfully downloaded {modelType} model to {modelPath}");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Failed to download {modelType} model: {ex.Message}");
-                    // Continue with other models even if one fails
-                }
-            }
-
-            Console.WriteLine("Model download process completed.");
-        }
-
-        /// <summary>
-        /// Downloads specific models for build/deployment purposes
-        /// This method is not restricted to debug mode and can be used in CI/CD pipelines
-        /// </summary>
-        /// <param name="modelTypes">Array of model types to download. If null, downloads base and large models.</param>
-        /// <param name="targetDirectory">Target directory for models. If null, uses default Models directory.</param>
-        /// <returns>Task representing the download operation</returns>
-        public async Task DownloadModelsForBuildAsync(string[] modelTypes = null, string targetDirectory = null)
-        {
-            // Default to essential models if none specified
-            modelTypes = modelTypes ?? new[] { "base", "large" };
-
-            // Use provided directory or default to Models subdirectory
-            string modelsDirectory = targetDirectory ?? Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Models");
-
-            // Ensure the Models directory exists
-            Directory.CreateDirectory(modelsDirectory);
-
-            Console.WriteLine($"Downloading {modelTypes.Length} model(s) to: {modelsDirectory}");
-
-            var downloader = new WhisperGgmlDownloader(_httpClient);
-
-            foreach (string modelType in modelTypes)
-            {
-                try
-                {
-                    string fileName = $"ggml-{modelType}.bin";
-                    string modelPath = Path.Combine(modelsDirectory, fileName);
-
-                    // Skip if model already exists
-                    if (File.Exists(modelPath))
-                    {
-                        Console.WriteLine($"Model {modelType} already exists at {modelPath}");
-                        continue;
-                    }
-
-                    Console.WriteLine($"Downloading {modelType} model...");
-
-                    // Download the model
-                    using (var modelStream = await downloader.GetGgmlModelAsync(GetGgmlType(modelType)))
-                    using (var fileStream = File.Create(modelPath))
-                    {
-                        await modelStream.CopyToAsync(fileStream);
-                    }
-
-                    Console.WriteLine($"Successfully downloaded {modelType} model to {modelPath}");
-
-                    // Log file size for verification
-                    var fileInfo = new FileInfo(modelPath);
-                    Console.WriteLine($"Model {modelType} size: {fileInfo.Length / (1024 * 1024):F1} MB");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Failed to download {modelType} model: {ex.Message}");
-                    throw new Exception($"Critical failure downloading model {modelType}: {ex.Message}", ex);
-                }
-            }
-
-            Console.WriteLine("Build model download process completed successfully.");
-        }
-
-        /// <summary>
-        /// Downloads models using direct HTTP download for CI/CD environments
-        /// This is a backup method if the Whisper.net downloader doesn't work in CI
-        /// </summary>
-        /// <param name="modelTypes">Array of model types to download</param>
-        /// <param name="targetDirectory">Target directory for models</param>
-        /// <returns>Task representing the download operation</returns>
-        public async Task DownloadModelsDirectAsync(string[] modelTypes = null, string targetDirectory = null)
-        {
-            modelTypes = modelTypes ?? new[] { "base", "large" };
-            string modelsDirectory = targetDirectory ?? Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Models");
-
-            Directory.CreateDirectory(modelsDirectory);
-
-            // HuggingFace direct download URLs
-            const string baseUrl = "https://huggingface.co/ggerganov/whisper.cpp/resolve/main";
-
-            Console.WriteLine($"Direct downloading {modelTypes.Length} model(s) to: {modelsDirectory}");
-
-            foreach (string modelType in modelTypes)
-            {
-                try
-                {
-                    string fileName = $"ggml-{modelType}.bin";
-                    string modelPath = Path.Combine(modelsDirectory, fileName);
-                    string downloadUrl = $"{baseUrl}/{fileName}";
-
-                    if (File.Exists(modelPath))
-                    {
-                        Console.WriteLine($"Model {modelType} already exists at {modelPath}");
-                        continue;
-                    }
-
-                    Console.WriteLine($"Direct downloading {modelType} model from {downloadUrl}...");
-
-                    using (var response = await _httpClient.GetAsync(downloadUrl))
-                    {
-                        response.EnsureSuccessStatusCode();
-
-                        using (var fileStream = File.Create(modelPath))
-                        {
-                            await response.Content.CopyToAsync(fileStream);
-                        }
-                    }
-
-                    var fileInfo = new FileInfo(modelPath);
-                    Console.WriteLine($"Successfully downloaded {modelType} model ({fileInfo.Length / (1024 * 1024):F1} MB)");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Failed to directly download {modelType} model: {ex.Message}");
-                    throw new Exception($"Critical failure downloading model {modelType}: {ex.Message}", ex);
-                }
-            }
-
-            Console.WriteLine("Direct model download process completed successfully.");
-        }
-
-        /// <summary>
         /// Ensures the model exists, first checking local content files, then downloading if necessary
         /// </summary>
         /// <param name="modelPath">Path where the model should be located</param>
@@ -270,54 +97,11 @@ namespace ForensicWhisperDeskZH.Transcription
         }
 
         /// <summary>
-        /// Gets a list of available local models
-        /// </summary>
-        /// <returns>Array of available model types</returns>
-        public string[] GetAvailableLocalModels()
-        {
-            var availableModels = new List<string>();
-            string baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
-            string modelsDirectory = Path.Combine(baseDirectory, "Models");
-
-            if (Directory.Exists(modelsDirectory))
-            {
-                var modelFiles = Directory.GetFiles(modelsDirectory, "ggml-*.bin");
-                foreach (var file in modelFiles)
-                {
-                    var fileName = Path.GetFileName(file);
-                    if (fileName.StartsWith("ggml-") && fileName.EndsWith(".bin"))
-                    {
-                        // Extract model type from filename
-                        var modelType = fileName.Substring("ggml-".Length, fileName.Length - "ggml-.bin".Length);
-                        availableModels.Add(modelType);
-                    }
-                }
-            }
-
-            return availableModels.ToArray();
-        }
-
-        /// <summary>
         /// Creates a WhisperFactory from the specified model path
         /// </summary>
         public WhisperFactory CreateFactory(string modelPath)
         {
             return WhisperFactory.FromPath(modelPath);
-        }
-
-        private static GgmlType GetGgmlType(string modelType)
-        {
-            return modelType.ToLowerInvariant() switch
-            {
-                "tiny" => GgmlType.Tiny,
-                "base" => GgmlType.Base,
-                "small" => GgmlType.Small,
-                "medium" => GgmlType.Medium,
-                "large" => GgmlType.LargeV3,
-                "turbo" => GgmlType.LargeV3Turbo,
-
-                _ => throw new ArgumentException($"Unknown model type: {modelType}"),
-            };
         }
 
         private static string GetStringFromGgmlType(GgmlType modelType)
