+++ ForensicWhisperDeskZH/app.config
@@ -1,17 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
     <configSections>
-        <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="ForensicWhisperDeskZH.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
-        </sectionGroup>
     </configSections>
-    <applicationSettings>
-      <ForensicWhisperDeskZH.Properties.Settings>
-            <setting name="ServerURI" serializeAs="String">
-                <value>localhost:8000/</value>
-            </setting>
-        </ForensicWhisperDeskZH.Properties.Settings>
-    </applicationSettings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
